# Conflicts:
#	data/src/main/java/io/gnosis/data/models/ChainInfo.kt

+++ app/src/test/java/io/gnosis/safe/ui/transactions/TransactionListViewModelTest.kt
@@ -1217,7 +1217,7 @@ class TransactionListViewModelTest {
     companion object {
 
         private val CHAIN = Chain(
-            BuildConfig.CHAIN_ID,
+            BuildConfig.CHAIN_ID.toBigInteger(),
             BuildConfig.BLOCKCHAIN_NAME,
             BuildConfig.CHAIN_TEXT_COLOR,
             BuildConfig.CHAIN_BACKGROUND_COLOR,
